requirements.txt
kubernetes==12.0.1
python-dotenv==0.15.0